--- MarathonRunner.AzureFunctions/Properties/ServiceDependencies/MarathonRunnerFunction - Zip Deploy/storage1.arm.json
@@ -1,70 +1,0 @@
-{
-  "$schema": "https://schema.management.azure.com/schemas/2018-05-01/subscriptionDeploymentTemplate.json#",
-  "contentVersion": "1.0.0.0",
-  "parameters": {
-    "resourceGroupName": {
-      "type": "string",
-      "defaultValue": "marathon-runner",
-      "metadata": {
-        "_parameterType": "resourceGroup",
-        "description": "このリソースのリソース グループの名前。追跡機能を向上させるため、同じリソース グループにリソースを配置することをお勧めします。"
-      }
-    },
-    "resourceGroupLocation": {
-      "type": "string",
-      "defaultValue": "japaneast",
-      "metadata": {
-        "_parameterType": "location",
-        "description": "リソース グループの場所。リソース グループには、リソースとは異なる場所を指定できます。"
-      }
-    },
-    "resourceLocation": {
-      "type": "string",
-      "defaultValue": "[parameters('resourceGroupLocation')]",
-      "metadata": {
-        "_parameterType": "location",
-        "description": "リソースの場所。既定では、リソース グループの場所が使用されます (リソース プロバイダーがその場所でサポートされていない場合を除く)。"
-      }
-    }
-  },
-  "resources": [
-    {
-      "type": "Microsoft.Resources/resourceGroups",
-      "name": "[parameters('resourceGroupName')]",
-      "location": "[parameters('resourceGroupLocation')]",
-      "apiVersion": "2019-10-01"
-    },
-    {
-      "type": "Microsoft.Resources/deployments",
-      "name": "[concat(parameters('resourceGroupName'), 'Deployment', uniqueString(concat('marathonrunnerstorage', subscription().subscriptionId)))]",
-      "resourceGroup": "[parameters('resourceGroupName')]",
-      "apiVersion": "2019-10-01",
-      "dependsOn": [
-        "[parameters('resourceGroupName')]"
-      ],
-      "properties": {
-        "mode": "Incremental",
-        "template": {
-          "$schema": "https://schema.management.azure.com/schemas/2019-04-01/deploymentTemplate.json#",
-          "contentVersion": "1.0.0.0",
-          "resources": [
-            {
-              "sku": {
-                "name": "Standard_LRS",
-                "tier": "Standard"
-              },
-              "kind": "Storage",
-              "name": "marathonrunnerstorage",
-              "type": "Microsoft.Storage/storageAccounts",
-              "location": "[parameters('resourceLocation')]",
-              "apiVersion": "2017-10-01"
-            }
-          ]
-        }
-      }
-    }
-  ],
-  "metadata": {
-    "_dependencyType": "storage.azure"
-  }
-}
