--- MarathonRunner.AzureFunctions/Properties/serviceDependencies.MarathonRunnerFunction.json
@@ -1,18 +1,0 @@
-{
-  "dependencies": {
-    "appInsights1": {
-      "secretStore": "AzureAppSettings",
-      "resourceId": "/subscriptions/[parameters('subscriptionId')]/resourceGroups/[parameters('resourceGroupName')]/providers/microsoft.insights/components/MarathonRunnerApplicationInsights",
-      "type": "appInsights.azure",
-      "connectionId": "APPLICATIONINSIGHTS_CONNECTION_STRING",
-      "dynamicId": null
-    },
-    "storage1": {
-      "secretStore": "AzureAppSettings",
-      "resourceId": "/subscriptions/[parameters('subscriptionId')]/resourceGroups/[parameters('resourceGroupName')]/providers/Microsoft.Storage/storageAccounts/marathonrunnerstorage",
-      "type": "storage.azure",
-      "connectionId": "AzureWebJobsStorage",
-      "dynamicId": null
-    }
-  }
-}
