--- MarathonRunner.AzureFunctions/Properties/serviceDependencies.json
@@ -1,14 +1,0 @@
-{
-  "dependencies": {
-    "appInsights1": {
-      "type": "appInsights",
-      "connectionId": "APPLICATIONINSIGHTS_CONNECTION_STRING",
-      "dynamicId": null
-    },
-    "storage1": {
-      "type": "storage",
-      "connectionId": "AzureWebJobsStorage",
-      "dynamicId": null
-    }
-  }
-}
