--- MarathonRunner.AzureFunctions/Properties/serviceDependencies.local.json
@@ -1,8 +1,0 @@
-{
-  "dependencies": {
-    "storage1": {
-      "type": "storage.emulator",
-      "connectionId": "AzureWebJobsStorage"
-    }
-  }
-}
