--- MarathonRunner.AzureFunctions/host.json
@@ -1,19 +1,0 @@
-{
-  "version": "2.0",
-  "logging": {
-    "applicationInsights": {
-      "samplingSettings": {
-        "isEnabled": true,
-        "excludedTypes": "Request"
-      }
-    }
-  },
-  "extensions": {
-    "http": {
-      "routePrefix": "api",
-      "maxOutstandingRequests": 3,
-      "maxConcurrentRequests": 1,
-      "dynamicThrottlesEnabled": true
-    }
-  }
-}
