+++ MarathonRunner.CloudRun/Dockerfile
@@ -1,0 +1,21 @@
+#See https://aka.ms/containerfastmode to understand how Visual Studio uses this Dockerfile to build your images for faster debugging.
+
+FROM mcr.microsoft.com/dotnet/aspnet:6.0 AS base
+WORKDIR /app
+EXPOSE 8080
+
+FROM mcr.microsoft.com/dotnet/sdk:6.0 AS build
+WORKDIR /src
+COPY ["MarathonRunner.CloudRun/MarathonRunner.CloudRun.csproj", "MarathonRunner.CloudRun/"]
+RUN dotnet restore "MarathonRunner.CloudRun/MarathonRunner.CloudRun.csproj"
+COPY . .
+WORKDIR "/src/MarathonRunner.CloudRun"
+RUN dotnet build "MarathonRunner.CloudRun.csproj" -c Release -o /app/build
+
+FROM build AS publish
+RUN dotnet publish "MarathonRunner.CloudRun.csproj" -c Release -o /app/publish
+
+FROM base AS final
+WORKDIR /app
+COPY --from=publish /app/publish .
+ENTRYPOINT ["dotnet", "MarathonRunner.CloudRun.dll"]
