+++ MarathonRunner.sln
@@ -9,7 +9,7 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MarathonRunner.Infrastructu
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MarathonRunner.Console", "MarathonRunner.Console\MarathonRunner.Console.csproj", "{1EDC7251-0DA0-4FF6-B6B2-C4E9074FEB7B}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MarathonRunner.AzureFunctions", "MarathonRunner.AzureFunctions\MarathonRunner.AzureFunctions.csproj", "{402F90DA-24FC-440A-B588-B7AC1B8013EA}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MarathonRunner.CloudRun", "MarathonRunner.CloudRun\MarathonRunner.CloudRun.csproj", "{9BCB6EC1-8A4F-4EF4-A6B4-23C45D518606}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -29,10 +29,10 @@ Global
 		{1EDC7251-0DA0-4FF6-B6B2-C4E9074FEB7B}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{1EDC7251-0DA0-4FF6-B6B2-C4E9074FEB7B}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{1EDC7251-0DA0-4FF6-B6B2-C4E9074FEB7B}.Release|Any CPU.Build.0 = Release|Any CPU
-		{402F90DA-24FC-440A-B588-B7AC1B8013EA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{402F90DA-24FC-440A-B588-B7AC1B8013EA}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{402F90DA-24FC-440A-B588-B7AC1B8013EA}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{402F90DA-24FC-440A-B588-B7AC1B8013EA}.Release|Any CPU.Build.0 = Release|Any CPU
+		{9BCB6EC1-8A4F-4EF4-A6B4-23C45D518606}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{9BCB6EC1-8A4F-4EF4-A6B4-23C45D518606}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{9BCB6EC1-8A4F-4EF4-A6B4-23C45D518606}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{9BCB6EC1-8A4F-4EF4-A6B4-23C45D518606}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
